Move ActivityButton type into its own submodule
